refactor(Light List): scene의 멤버 변수 type을 Actor로 바꿈
- Outliner Gui에 Light를 추가하기 위해 Scene에서 m_lights를
- array LightActor로 사용하던걸 vector<Actor>로 Refactoring

+++ DXEngine/LightActor.cpp
@@ -65,7 +65,7 @@ namespace DE {
 			return; // ºûÀÌ ²¨Á®ÀÖÀ¸¸é ·»´õ¸µÇÏÁö ¾ÊÀ½
 
 		if (m_light.type & LIGHT_SHADOW) {
-			GET_SINGLE(RenderBase)->SetShadowViewport(m_shadowWidth, m_shadowHeight);
+			GET_SINGLE(RenderBase)->SetShadowViewport((float)m_shadowWidth, (float)m_shadowHeight);
 			GET_SINGLE(RenderBase)->SetShadowMap(m_lightID);
 			SetGlobals(m_shadowGlobalConsts.Get());
